requirements.txt
Pillow
python-dotenv
selenium